python-backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
strawberry-graphql[fastapi]==0.217.0

# Database
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI Integration
openai==1.10.0
anthropic==0.7.8
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2